+++ pkg/analyzer/test/src/dart/element/element_test.dart
@@ -204,7 +204,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     String methodName = "m";
     MethodElement method = ElementFactory.methodElement(methodName, intNone);
     classA.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpConcreteMethod(methodName, library), same(method));
   }
@@ -220,7 +220,7 @@ class ClassElementImplTest extends AbstractTypeTest {
         ElementFactory.methodElement(methodName, intNone);
     method.isAbstract = true;
     classA.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpConcreteMethod(methodName, library), isNull);
   }
@@ -244,7 +244,7 @@ class ClassElementImplTest extends AbstractTypeTest {
         ElementFactory.methodElement(methodName, intNone);
     method.isAbstract = true;
     classB.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpConcreteMethod(methodName, library),
         same(inheritedMethod));
@@ -267,7 +267,7 @@ class ClassElementImplTest extends AbstractTypeTest {
         ElementFactory.classElement("B", interfaceTypeStar(classA));
     MethodElement method = ElementFactory.methodElement(methodName, intNone);
     classB.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpConcreteMethod(methodName, library), same(method));
   }
@@ -291,7 +291,7 @@ class ClassElementImplTest extends AbstractTypeTest {
         ElementFactory.classElement("B", interfaceTypeStar(classA));
     MethodElement method = ElementFactory.methodElement(methodName, intNone);
     classB.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpConcreteMethod(methodName, library), same(method));
   }
@@ -310,7 +310,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     classA.methods = <MethodElement>[inheritedMethod];
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpConcreteMethod(methodName, library),
         same(inheritedMethod));
@@ -321,7 +321,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     // }
     LibraryElementImpl library = _newLibrary();
     var classA = class_(name: 'A');
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpConcreteMethod("m", library), isNull);
   }
@@ -336,7 +336,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     PropertyAccessorElement getter =
         ElementFactory.getterElement(getterName, false, intNone);
     classA.accessors = <PropertyAccessorElement>[getter];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpGetter(getterName, library), same(getter));
   }
@@ -355,7 +355,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     classA.accessors = <PropertyAccessorElement>[getter];
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpGetter(getterName, library), same(getter));
   }
@@ -365,7 +365,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     // }
     LibraryElementImpl library = _newLibrary();
     var classA = class_(name: 'A');
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpGetter("g", library), isNull);
   }
@@ -380,7 +380,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
     classA.supertype = interfaceTypeStar(classB);
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classA.lookUpGetter("g", library), isNull);
   }
@@ -395,7 +395,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     PropertyAccessorElement getter =
         ElementFactory.getterElement(getterName, false, intNone);
     classA.accessors = <PropertyAccessorElement>[getter];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpInheritedConcreteGetter(getterName, library), isNull);
   }
@@ -414,7 +414,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     classA.accessors = <PropertyAccessorElement>[inheritedGetter];
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpInheritedConcreteGetter(getterName, library),
         same(inheritedGetter));
@@ -425,7 +425,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     // }
     LibraryElementImpl library = _newLibrary();
     var classA = class_(name: 'A');
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpInheritedConcreteGetter("g", library), isNull);
   }
@@ -440,7 +440,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
     classA.supertype = interfaceTypeStar(classB);
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classA.lookUpInheritedConcreteGetter("g", library), isNull);
   }
@@ -454,7 +454,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     String methodName = "m";
     MethodElement method = ElementFactory.methodElement(methodName, intNone);
     classA.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpInheritedConcreteMethod(methodName, library), isNull);
   }
@@ -478,7 +478,7 @@ class ClassElementImplTest extends AbstractTypeTest {
         ElementFactory.methodElement(methodName, intNone);
     method.isAbstract = true;
     classB.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpInheritedConcreteMethod(methodName, library),
         same(inheritedMethod));
@@ -501,7 +501,7 @@ class ClassElementImplTest extends AbstractTypeTest {
         ElementFactory.classElement("B", interfaceTypeStar(classA));
     MethodElement method = ElementFactory.methodElement(methodName, intNone);
     classB.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpInheritedConcreteMethod(methodName, library),
         same(inheritedMethod));
@@ -526,7 +526,7 @@ class ClassElementImplTest extends AbstractTypeTest {
         ElementFactory.classElement("B", interfaceTypeStar(classA));
     MethodElement method = ElementFactory.methodElement(methodName, intNone);
     classB.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpInheritedConcreteMethod(methodName, library), isNull);
   }
@@ -559,7 +559,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     MethodElementImpl method =
         ElementFactory.methodElement(methodName, intNone);
     classC.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB, classC];
     expect(classC.lookUpInheritedConcreteMethod(methodName, library),
         same(inheritedMethod));
@@ -579,7 +579,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     classA.methods = <MethodElement>[inheritedMethod];
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpInheritedConcreteMethod(methodName, library),
         same(inheritedMethod));
@@ -590,7 +590,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     // }
     LibraryElementImpl library = _newLibrary();
     var classA = class_(name: 'A');
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpInheritedConcreteMethod("m", library), isNull);
   }
@@ -605,7 +605,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     PropertyAccessorElement setter =
         ElementFactory.setterElement(setterName, false, intNone);
     classA.accessors = <PropertyAccessorElement>[setter];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpInheritedConcreteSetter(setterName, library), isNull);
   }
@@ -624,7 +624,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     classA.accessors = <PropertyAccessorElement>[setter];
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpInheritedConcreteSetter(setterName, library),
         same(setter));
@@ -635,7 +635,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     // }
     LibraryElementImpl library = _newLibrary();
     var classA = class_(name: 'A');
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpInheritedConcreteSetter("s", library), isNull);
   }
@@ -650,7 +650,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
     classA.supertype = interfaceTypeStar(classB);
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classA.lookUpInheritedConcreteSetter("s", library), isNull);
   }
@@ -664,7 +664,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     String methodName = "m";
     MethodElement method = ElementFactory.methodElement(methodName, intNone);
     classA.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpInheritedMethod(methodName, library), isNull);
   }
@@ -686,7 +686,7 @@ class ClassElementImplTest extends AbstractTypeTest {
         ElementFactory.classElement("B", interfaceTypeStar(classA));
     MethodElement method = ElementFactory.methodElement(methodName, intNone);
     classB.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpInheritedMethod(methodName, library),
         same(inheritedMethod));
@@ -706,7 +706,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     classA.methods = <MethodElement>[inheritedMethod];
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpInheritedMethod(methodName, library),
         same(inheritedMethod));
@@ -717,7 +717,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     // }
     LibraryElementImpl library = _newLibrary();
     var classA = class_(name: 'A');
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpInheritedMethod("m", library), isNull);
   }
@@ -728,7 +728,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     String methodName = "m";
     MethodElement method = ElementFactory.methodElement(methodName, intNone);
     classA.methods = <MethodElement>[method];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpMethod(methodName, library), same(method));
   }
@@ -741,7 +741,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     classA.methods = <MethodElement>[method];
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpMethod(methodName, library), same(method));
   }
@@ -749,7 +749,7 @@ class ClassElementImplTest extends AbstractTypeTest {
   void test_lookUpMethod_undeclared() {
     LibraryElementImpl library = _newLibrary();
     var classA = class_(name: 'A');
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpMethod("m", library), isNull);
   }
@@ -760,7 +760,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
     classA.supertype = interfaceTypeStar(classB);
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classA.lookUpMethod("m", library), isNull);
   }
@@ -775,7 +775,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     PropertyAccessorElement setter =
         ElementFactory.setterElement(setterName, false, intNone);
     classA.accessors = <PropertyAccessorElement>[setter];
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpSetter(setterName, library), same(setter));
   }
@@ -794,7 +794,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     classA.accessors = <PropertyAccessorElement>[setter];
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classB.lookUpSetter(setterName, library), same(setter));
   }
@@ -804,7 +804,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     // }
     LibraryElementImpl library = _newLibrary();
     var classA = class_(name: 'A');
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA];
     expect(classA.lookUpSetter("s", library), isNull);
   }
@@ -819,7 +819,7 @@ class ClassElementImplTest extends AbstractTypeTest {
     ClassElementImpl classB =
         ElementFactory.classElement("B", interfaceTypeStar(classA));
     classA.supertype = interfaceTypeStar(classB);
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classA, classB];
     expect(classA.lookUpSetter("s", library), isNull);
   }
@@ -835,7 +835,7 @@ class CompilationUnitElementImplTest {
         ElementFactory.compilationUnit("/lib.dart");
     String className = "C";
     ClassElement classElement = ElementFactory.classElement2(className);
-    unit.types = <ClassElement>[classElement];
+    unit.classes = <ClassElement>[classElement];
     expect(unit.getType(className), same(classElement));
   }
 
@@ -844,7 +844,7 @@ class CompilationUnitElementImplTest {
         ElementFactory.compilationUnit("/lib.dart");
     String className = "C";
     ClassElement classElement = ElementFactory.classElement2(className);
-    unit.types = <ClassElement>[classElement];
+    unit.classes = <ClassElement>[classElement];
     expect(unit.getType("${className}x"), isNull);
   }
 }
@@ -881,7 +881,7 @@ class ElementImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     ClassElementImpl classElement = ElementFactory.classElement2("C");
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classElement];
     FieldElement field = ElementFactory.fieldElement(
       "next",
@@ -906,7 +906,7 @@ class ElementImplTest extends AbstractTypeTest {
     LibraryElementImpl library1 =
         ElementFactory.library(_analysisContext, "lib1");
     ClassElement classElement = ElementFactory.classElement2("_C");
-    (library1.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library1.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classElement];
     LibraryElementImpl library2 =
         ElementFactory.library(_analysisContext, "lib2");
@@ -917,7 +917,7 @@ class ElementImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     ClassElement classElement = ElementFactory.classElement2("_C");
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classElement];
     expect(classElement.isAccessibleIn(library), isTrue);
   }
@@ -926,7 +926,7 @@ class ElementImplTest extends AbstractTypeTest {
     LibraryElementImpl library1 =
         ElementFactory.library(_analysisContext, "lib1");
     ClassElement classElement = ElementFactory.classElement2("C");
-    (library1.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library1.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classElement];
     LibraryElementImpl library2 =
         ElementFactory.library(_analysisContext, "lib2");
@@ -937,7 +937,7 @@ class ElementImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     ClassElement classElement = ElementFactory.classElement2("C");
-    (library.definingCompilationUnit as CompilationUnitElementImpl).types =
+    (library.definingCompilationUnit as CompilationUnitElementImpl).classes =
         <ClassElement>[classElement];
     expect(classElement.isAccessibleIn(library), isTrue);
   }
@@ -1650,7 +1650,7 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[classA];
     expect(typeA.lookUpGetter(getterName, library), same(getterG));
   }
 
@@ -1671,7 +1671,10 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA, classB];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[
+      classA,
+      classB
+    ];
     expect(typeB.lookUpGetter(getterName, library), same(getterG));
   }
 
@@ -1703,7 +1706,7 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     var unit = library.definingCompilationUnit as CompilationUnitElementImpl;
-    unit.types = <ClassElement>[classB, classM1, classM2, classC];
+    unit.classes = <ClassElement>[classB, classM1, classM2, classC];
     expect(
         interfaceTypeStar(classC).lookUpGetter(getterName, library), getterM2g);
   }
@@ -1721,7 +1724,10 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA, classB];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[
+      classA,
+      classB
+    ];
     expect(typeA.lookUpGetter("g", library), isNull);
   }
 
@@ -1735,7 +1741,7 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[classA];
     expect(typeA.lookUpGetter("g", library), isNull);
   }
 
@@ -1753,7 +1759,7 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[classA];
     expect(typeA.lookUpMethod(methodName, library), same(methodM));
   }
 
@@ -1774,7 +1780,10 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA, classB];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[
+      classA,
+      classB
+    ];
     expect(typeB.lookUpMethod(methodName, library), same(methodM));
   }
 
@@ -1805,7 +1814,7 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     var unit = library.definingCompilationUnit as CompilationUnitElementImpl;
-    unit.types = <ClassElement>[classB, classM1, classM2, classC];
+    unit.classes = <ClassElement>[classB, classM1, classM2, classC];
     expect(
         interfaceTypeStar(classC).lookUpMethod(methodName, library), methodM2m);
   }
@@ -1836,7 +1845,7 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[A];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[A];
     //
     // B<I>
     //
@@ -1864,7 +1873,10 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA, classB];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[
+      classA,
+      classB
+    ];
     expect(typeA.lookUpMethod("m", library), isNull);
   }
 
@@ -1878,7 +1890,7 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[classA];
     expect(typeA.lookUpMethod("m", library), isNull);
   }
 
@@ -1896,7 +1908,7 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[classA];
     expect(typeA.lookUpSetter(setterName, library), same(setterS));
   }
 
@@ -1917,7 +1929,10 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA, classB];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[
+      classA,
+      classB
+    ];
     expect(typeB.lookUpSetter(setterName, library), same(setterS));
   }
 
@@ -1949,7 +1964,7 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     var unit = library.definingCompilationUnit as CompilationUnitElementImpl;
-    unit.types = <ClassElement>[classB, classM1, classM2, classC];
+    unit.classes = <ClassElement>[classB, classM1, classM2, classC];
     expect(
         interfaceTypeStar(classC).lookUpGetter(setterName, library), setterM2g);
   }
@@ -1967,7 +1982,10 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA, classB];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[
+      classA,
+      classB
+    ];
     expect(typeA.lookUpSetter("s", library), isNull);
   }
 
@@ -1981,7 +1999,7 @@ class InterfaceTypeImplTest extends AbstractTypeTest {
     LibraryElementImpl library =
         ElementFactory.library(_analysisContext, "lib");
     CompilationUnitElement unit = library.definingCompilationUnit;
-    (unit as CompilationUnitElementImpl).types = <ClassElement>[classA];
+    (unit as CompilationUnitElementImpl).classes = <ClassElement>[classA];
     expect(typeA.lookUpSetter("s", library), isNull);
   }