+++ pkg/analyzer/CHANGELOG.md
@@ -3,6 +3,7 @@
   explicitly  that there are always (possibly empty) strings as the first
   and the last elements of an interpolation.
 * Deprecated `ImportElement.prefixOffset`, use `prefix.nameOffset` instead.
+* Deprecated `CompilationUnitElement.types`, use `classes` instead.
 
 ## 1.7.0
 * Require `meta: ^1.4.0`.